Initial refactoring of the GhidraMCP project code's basic structure, separating service logic related to different web paths and performing preliminary classification based on service functionality types.
Additionally, I changed pom.xml so now it supports importing other packages from Maven, which will be compiled into `GhidraMCP.jar`.

I think it's a promising tool and intend to contribute more soon. If possible, you can build a frontend to make users possible to choose routes to use as they wish to save tokens, like Yakit :)
