+++ src/main/java/com/lauriewired/handlers/act/DecompileFunctionByAddress.java
@@ -1,0 +1,56 @@
+package com.lauriewired.handlers.act;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.app.decompiler.DecompInterface;
+import ghidra.app.decompiler.DecompileResults;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.address.Address;
+import ghidra.program.model.listing.Function;
+import ghidra.program.model.listing.Program;
+import ghidra.util.task.ConsoleTaskMonitor;
+
+import java.io.IOException;
+import java.util.Map;
+
+import static com.lauriewired.util.ParseUtils.parseQueryParams;
+import static com.lauriewired.util.ParseUtils.sendResponse;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class DecompileFunctionByAddress extends Handler {
+    public DecompileFunctionByAddress(PluginTool tool) {
+        super(tool, "/decompile_function");
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws IOException {
+        Map<String, String> qparams = parseQueryParams(exchange);
+        String address = qparams.get("address");
+        sendResponse(exchange, decompileFunctionByAddress(address));
+    }
+
+    /**
+     * Decompile a function at the given address
+     */
+    private String decompileFunctionByAddress(String addressStr) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return "No program loaded";
+        if (addressStr == null || addressStr.isEmpty()) return "Address is required";
+
+        try {
+            Address addr = program.getAddressFactory().getAddress(addressStr);
+            Function func = program.getListing().getFunctionContaining(addr);
+            if (func == null) return "No function found at or containing address " + addressStr;
+
+            DecompInterface decomp = new DecompInterface();
+            decomp.openProgram(program);
+            DecompileResults result = decomp.decompileFunction(func, 30, new ConsoleTaskMonitor());
+
+            return (result != null && result.decompileCompleted())
+                    ? result.getDecompiledFunction().getC()
+                    : "Decompilation failed";
+        } catch (Exception e) {
+            return "Error decompiling function: " + e.getMessage();
+        }
+    }
+}
