+++ src/main/java/com/lauriewired/handlers/act/DisassembleFunction.java
@@ -1,0 +1,64 @@
+package com.lauriewired.handlers.act;
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.address.Address;
+import ghidra.program.model.listing.*;
+
+import java.io.IOException;
+import java.util.Map;
+
+import static com.lauriewired.util.ParseUtils.parseQueryParams;
+import static com.lauriewired.util.ParseUtils.sendResponse;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class DisassembleFunction extends Handler {
+    public DisassembleFunction(PluginTool tool) {
+        super(tool, "/disassemble_function");
+    }
+
+    public void handle(HttpExchange exchange) throws IOException {
+        Map<String, String> qparams = parseQueryParams(exchange);
+        String address = qparams.get("address");
+        sendResponse(exchange, disassembleFunction(address));
+    }
+
+    /**
+     * Get assembly code for a function
+     */
+    private String disassembleFunction(String addressStr) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return "No program loaded";
+        if (addressStr == null || addressStr.isEmpty()) return "Address is required";
+
+        try {
+            Address addr = program.getAddressFactory().getAddress(addressStr);
+            Function func = program.getListing().getFunctionContaining(addr);
+            if (func == null) return "No function found at or containing address " + addressStr;
+
+            StringBuilder result = new StringBuilder();
+            Listing listing = program.getListing();
+            Address start = func.getEntryPoint();
+            Address end = func.getBody().getMaxAddress();
+
+            InstructionIterator instructions = listing.getInstructions(start, true);
+            while (instructions.hasNext()) {
+                Instruction instr = instructions.next();
+                if (instr.getAddress().compareTo(end) > 0) {
+                    break; // Stop if we've gone past the end of the function
+                }
+                String comment = listing.getComment(CodeUnit.EOL_COMMENT, instr.getAddress());
+                comment = (comment != null) ? "; " + comment : "";
+
+                result.append(String.format("%s: %s %s\n",
+                        instr.getAddress(),
+                        instr.toString(),
+                        comment));
+            }
+
+            return result.toString();
+        } catch (Exception e) {
+            return "Error disassembling function: " + e.getMessage();
+        }
+    }
+}
