+++ src/main/java/com/lauriewired/handlers/get/GetAllClassNames.java
@@ -1,0 +1,45 @@
+package com.lauriewired.handlers.get;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.listing.Program;
+import ghidra.program.model.symbol.Namespace;
+import ghidra.program.model.symbol.Symbol;
+
+import java.io.IOException;
+import java.util.*;
+
+import static com.lauriewired.util.ParseUtils.*;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class GetAllClassNames extends Handler {
+    public GetAllClassNames(PluginTool tool) {
+        super(tool, "/classes");
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws IOException {
+        Map<String, String> qparams = parseQueryParams(exchange);
+        int offset = parseIntOrDefault(qparams.get("offset"), 0);
+        int limit  = parseIntOrDefault(qparams.get("limit"),  100);
+        sendResponse(exchange, generateResponse(offset, limit));
+    }
+
+    private String generateResponse(int offset, int limit) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return "No program loaded";
+
+        Set<String> classNames = new HashSet<>();
+        for (Symbol symbol : program.getSymbolTable().getAllSymbols(true)) {
+            Namespace ns = symbol.getParentNamespace();
+            if (ns != null && !ns.isGlobal()) {
+                classNames.add(ns.getName());
+            }
+        }
+        // Convert set to list for pagination
+        List<String> sorted = new ArrayList<>(classNames);
+        Collections.sort(sorted);
+        return paginateList(sorted, offset, limit);
+    }
+}
