+++ src/main/java/com/lauriewired/handlers/get/GetCurrentFunction.java
@@ -1,0 +1,47 @@
+package com.lauriewired.handlers.get;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.app.services.CodeViewerService;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.listing.Function;
+import ghidra.program.model.listing.Program;
+import ghidra.program.util.ProgramLocation;
+
+import java.io.IOException;
+
+import static com.lauriewired.util.ParseUtils.sendResponse;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class GetCurrentFunction extends Handler {
+    public GetCurrentFunction(PluginTool tool) {
+        super(tool, "/get_current_function");
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws IOException {
+        sendResponse(exchange, getCurrentFunction());
+    }
+
+    /**
+     * Get current function selected in Ghidra GUI
+     */
+    private String getCurrentFunction() {
+        CodeViewerService service = tool.getService(CodeViewerService.class);
+        if (service == null) return "Code viewer service not available";
+
+        ProgramLocation location = service.getCurrentLocation();
+        if (location == null) return "No current location";
+
+        Program program = getCurrentProgram(tool);
+        if (program == null) return "No program loaded";
+
+        Function func = program.getFunctionManager().getFunctionContaining(location.getAddress());
+        if (func == null) return "No function at current location: " + location.getAddress();
+
+        return String.format("Function: %s at %s\nSignature: %s",
+                func.getName(),
+                func.getEntryPoint(),
+                func.getSignature());
+    }
+}
