+++ src/main/java/com/lauriewired/handlers/get/GetFunctionXrefs.java
@@ -1,0 +1,73 @@
+package com.lauriewired.handlers.get;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.address.Address;
+import ghidra.program.model.listing.Function;
+import ghidra.program.model.listing.FunctionManager;
+import ghidra.program.model.listing.Program;
+import ghidra.program.model.symbol.RefType;
+import ghidra.program.model.symbol.Reference;
+import ghidra.program.model.symbol.ReferenceIterator;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import static com.lauriewired.util.ParseUtils.*;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class GetFunctionXrefs extends Handler {
+    public GetFunctionXrefs(PluginTool tool) {
+        super(tool, "/function_xrefs");
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws Exception {
+        Map<String, String> qparams = parseQueryParams(exchange);
+        String name = qparams.get("name");
+        int offset = parseIntOrDefault(qparams.get("offset"), 0);
+        int limit = parseIntOrDefault(qparams.get("limit"), 100);
+        sendResponse(exchange, getFunctionXrefs(name, offset, limit));
+    }
+
+    /**
+     * Get all references to a specific function by name
+     */
+    private String getFunctionXrefs(String functionName, int offset, int limit) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return "No program loaded";
+        if (functionName == null || functionName.isEmpty()) return "Function name is required";
+
+        try {
+            List<String> refs = new ArrayList<>();
+            FunctionManager funcManager = program.getFunctionManager();
+            for (Function function : funcManager.getFunctions(true)) {
+                if (function.getName().equals(functionName)) {
+                    Address entryPoint = function.getEntryPoint();
+                    ReferenceIterator refIter = program.getReferenceManager().getReferencesTo(entryPoint);
+
+                    while (refIter.hasNext()) {
+                        Reference ref = refIter.next();
+                        Address fromAddr = ref.getFromAddress();
+                        RefType refType = ref.getReferenceType();
+
+                        Function fromFunc = funcManager.getFunctionContaining(fromAddr);
+                        String funcInfo = (fromFunc != null) ? " in " + fromFunc.getName() : "";
+
+                        refs.add(String.format("From %s%s [%s]", fromAddr, funcInfo, refType.getName()));
+                    }
+                }
+            }
+
+            if (refs.isEmpty()) {
+                return "No references found to function: " + functionName;
+            }
+
+            return paginateList(refs, offset, limit);
+        } catch (Exception e) {
+            return "Error getting function references: " + e.getMessage();
+        }
+    }
+}
