+++ src/main/java/com/lauriewired/handlers/get/ListDefinedData.java
@@ -1,0 +1,54 @@
+package com.lauriewired.handlers.get;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.listing.Data;
+import ghidra.program.model.listing.DataIterator;
+import ghidra.program.model.listing.Program;
+import ghidra.program.model.mem.MemoryBlock;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import static com.lauriewired.util.ParseUtils.*;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class ListDefinedData extends Handler {
+    public ListDefinedData(PluginTool tool) {
+        super(tool, "/data");
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws IOException {
+        Map<String, String> qparams = parseQueryParams(exchange);
+        int offset = parseIntOrDefault(qparams.get("offset"), 0);
+        int limit  = parseIntOrDefault(qparams.get("limit"),  100);
+        sendResponse(exchange, listDefinedData(offset, limit));
+    }
+
+    private String listDefinedData(int offset, int limit) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return "No program loaded";
+
+        List<String> lines = new ArrayList<>();
+        for (MemoryBlock block : program.getMemory().getBlocks()) {
+            DataIterator it = program.getListing().getDefinedData(block.getStart(), true);
+            while (it.hasNext()) {
+                Data data = it.next();
+                if (block.contains(data.getAddress())) {
+                    String label   = data.getLabel() != null ? data.getLabel() : "(unnamed)";
+                    String valRepr = data.getDefaultValueRepresentation();
+                    lines.add(String.format("%s: %s = %s",
+                            data.getAddress(),
+                            escapeNonAscii(label),
+                            escapeNonAscii(valRepr)
+                    ));
+                }
+            }
+        }
+        return paginateList(lines, offset, limit);
+    }
+}
