+++ src/main/java/com/lauriewired/handlers/get/ListDefinedStrings.java
@@ -1,0 +1,92 @@
+package com.lauriewired.handlers.get;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.data.DataType;
+import ghidra.program.model.listing.Data;
+import ghidra.program.model.listing.DataIterator;
+import ghidra.program.model.listing.Program;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import static com.lauriewired.util.ParseUtils.*;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class ListDefinedStrings extends Handler {
+    public ListDefinedStrings(PluginTool tool) {
+        super(tool, "/strings");
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws Exception {
+        Map<String, String> qparams = parseQueryParams(exchange);
+        int offset = parseIntOrDefault(qparams.get("offset"), 0);
+        int limit = parseIntOrDefault(qparams.get("limit"), 100);
+        String filter = qparams.get("filter");
+        sendResponse(exchange, listDefinedStrings(offset, limit, filter));
+    }
+
+    /**
+     * List all defined strings in the program with their addresses
+     */
+    private String listDefinedStrings(int offset, int limit, String filter) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return "No program loaded";
+
+        List<String> lines = new ArrayList<>();
+        DataIterator dataIt = program.getListing().getDefinedData(true);
+
+        while (dataIt.hasNext()) {
+            Data data = dataIt.next();
+
+            if (data != null && isStringData(data)) {
+                String value = data.getValue() != null ? data.getValue().toString() : "";
+
+                if (filter == null || value.toLowerCase().contains(filter.toLowerCase())) {
+                    String escapedValue = escapeString(value);
+                    lines.add(String.format("%s: \"%s\"", data.getAddress(), escapedValue));
+                }
+            }
+        }
+
+        return paginateList(lines, offset, limit);
+    }
+
+    /**
+     * Check if the given data is a string type
+     */
+    private boolean isStringData(Data data) {
+        if (data == null) return false;
+
+        DataType dt = data.getDataType();
+        String typeName = dt.getName().toLowerCase();
+        return typeName.contains("string") || typeName.contains("char") || typeName.equals("unicode");
+    }
+
+    /**
+     * Escape special characters in a string for display
+     */
+    private String escapeString(String input) {
+        if (input == null) return "";
+
+        StringBuilder sb = new StringBuilder();
+        for (int i = 0; i < input.length(); i++) {
+            char c = input.charAt(i);
+            if (c >= 32 && c < 127) {
+                sb.append(c);
+            } else if (c == '\n') {
+                sb.append("\\n");
+            } else if (c == '\r') {
+                sb.append("\\r");
+            } else if (c == '\t') {
+                sb.append("\\t");
+            } else {
+                sb.append(String.format("\\x%02x", (int)c & 0xFF));
+            }
+        }
+        return sb.toString();
+    }
+}
