+++ src/main/java/com/lauriewired/handlers/get/ListImports.java
@@ -1,0 +1,40 @@
+package com.lauriewired.handlers.get;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.listing.Program;
+import ghidra.program.model.symbol.Symbol;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+import static com.lauriewired.util.ParseUtils.*;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class ListImports extends Handler {
+    public ListImports(PluginTool tool) {
+        super(tool, "/imports");
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws IOException {
+        Map<String, String> qparams = parseQueryParams(exchange);
+        int offset = parseIntOrDefault(qparams.get("offset"), 0);
+        int limit  = parseIntOrDefault(qparams.get("limit"),  100);
+        sendResponse(exchange, listImports(offset, limit));
+    }
+
+    private String listImports(int offset, int limit) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return "No program loaded";
+
+        List<String> lines = new ArrayList<>();
+        for (Symbol symbol : program.getSymbolTable().getExternalSymbols()) {
+            lines.add(symbol.getName() + " -> " + symbol.getAddress());
+        }
+        return paginateList(lines, offset, limit);
+    }
+}
