+++ src/main/java/com/lauriewired/handlers/set/RenameFunction.java
@@ -1,0 +1,63 @@
+package com.lauriewired.handlers.set;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.listing.Function;
+import ghidra.program.model.listing.Program;
+import ghidra.program.model.symbol.SourceType;
+import ghidra.util.Msg;
+
+import javax.swing.*;
+import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+import java.util.Map;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+import static com.lauriewired.util.ParseUtils.parsePostParams;
+import static com.lauriewired.util.ParseUtils.sendResponse;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class RenameFunction extends Handler {
+    public RenameFunction(PluginTool tool) {
+        super(tool, "/renameFunction");
+    }
+
+    public void handle(HttpExchange exchange) throws IOException {
+        Map<String, String> params = parsePostParams(exchange);
+        String response = rename(params.get("oldName"), params.get("newName"))
+                ? "Renamed successfully" : "Rename failed";
+        sendResponse(exchange, response);
+    }
+
+    private boolean rename(String oldName, String newName) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return false;
+
+        AtomicBoolean successFlag = new AtomicBoolean(false);
+        try {
+            SwingUtilities.invokeAndWait(() -> {
+                int tx = program.startTransaction("Rename function via HTTP");
+                try {
+                    for (Function func : program.getFunctionManager().getFunctions(true)) {
+                        if (func.getName().equals(oldName)) {
+                            func.setName(newName, SourceType.USER_DEFINED);
+                            successFlag.set(true);
+                            break;
+                        }
+                    }
+                }
+                catch (Exception e) {
+                    Msg.error(this, "Error renaming function", e);
+                }
+                finally {
+                    program.endTransaction(tx, successFlag.get());
+                }
+            });
+        }
+        catch (InterruptedException | InvocationTargetException e) {
+            Msg.error(this, "Failed to execute rename on Swing thread", e);
+        }
+        return successFlag.get();
+    }
+}
