+++ src/main/java/com/lauriewired/handlers/set/RenameFunctionByAddress.java
@@ -1,0 +1,81 @@
+package com.lauriewired.handlers.set;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.address.Address;
+import ghidra.program.model.listing.Function;
+import ghidra.program.model.listing.Program;
+import ghidra.program.model.symbol.SourceType;
+import ghidra.util.Msg;
+
+import javax.swing.*;
+import java.lang.reflect.InvocationTargetException;
+import java.util.Map;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+import static com.lauriewired.util.ParseUtils.parsePostParams;
+import static com.lauriewired.util.ParseUtils.sendResponse;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class RenameFunctionByAddress extends Handler {
+    public RenameFunctionByAddress(PluginTool tool) {
+        super(tool, "/rename_function_by_address");
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws Exception {
+        Map<String, String> params = parsePostParams(exchange);
+        String functionAddress = params.get("function_address");
+        String newName = params.get("new_name");
+        boolean success = renameFunctionByAddress(functionAddress, newName);
+        sendResponse(exchange, success ? "Function renamed successfully" : "Failed to rename function");
+    }
+
+    /**
+     * Rename a function by its address
+     */
+    private boolean renameFunctionByAddress(String functionAddrStr, String newName) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return false;
+        if (functionAddrStr == null || functionAddrStr.isEmpty() ||
+                newName == null || newName.isEmpty()) {
+            return false;
+        }
+
+        AtomicBoolean success = new AtomicBoolean(false);
+
+        try {
+            SwingUtilities.invokeAndWait(() -> {
+                performFunctionRename(program, functionAddrStr, newName, success);
+            });
+        } catch (InterruptedException | InvocationTargetException e) {
+            Msg.error(this, "Failed to execute rename function on Swing thread", e);
+        }
+
+        return success.get();
+    }
+
+    /**
+     * Helper method to perform the actual function rename within a transaction
+     */
+    private void performFunctionRename(Program program, String functionAddrStr, String newName, AtomicBoolean success) {
+        int tx = program.startTransaction("Rename function by address");
+        try {
+            Address addr = program.getAddressFactory().getAddress(functionAddrStr);
+            Function func = program.getListing().getFunctionContaining(addr);
+
+            if (func == null) {
+                Msg.error(this, "Could not find function at address: " + functionAddrStr);
+                return;
+            }
+
+            func.setName(newName, SourceType.USER_DEFINED);
+            success.set(true);
+        } catch (Exception e) {
+            Msg.error(this, "Error renaming function by address", e);
+        } finally {
+            program.endTransaction(tx, success.get());
+        }
+    }
+}
