+++ src/main/java/com/lauriewired/handlers/set/SearchFunctions.java
@@ -1,0 +1,53 @@
+package com.lauriewired.handlers.set;
+
+import com.lauriewired.handlers.Handler;
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.framework.plugintool.PluginTool;
+import ghidra.program.model.listing.Function;
+import ghidra.program.model.listing.Program;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+
+import static com.lauriewired.util.ParseUtils.*;
+import static ghidra.program.util.GhidraProgramUtilities.getCurrentProgram;
+
+public final class SearchFunctions extends Handler {
+    public SearchFunctions(PluginTool tool) {
+        super(tool, "/searchFunctions");
+    }
+
+    @Override
+    public void handle(HttpExchange exchange) throws IOException {
+        Map<String, String> qparams = parseQueryParams(exchange);
+        String searchTerm = qparams.get("query");
+        int offset = parseIntOrDefault(qparams.get("offset"), 0);
+        int limit = parseIntOrDefault(qparams.get("limit"), 100);
+        sendResponse(exchange, searchFunctionsByName(searchTerm, offset, limit));
+    }
+
+    private String searchFunctionsByName(String searchTerm, int offset, int limit) {
+        Program program = getCurrentProgram(tool);
+        if (program == null) return "No program loaded";
+        if (searchTerm == null || searchTerm.isEmpty()) return "Search term is required";
+
+        List<String> matches = new ArrayList<>();
+        for (Function func : program.getFunctionManager().getFunctions(true)) {
+            String name = func.getName();
+            // simple substring match
+            if (name.toLowerCase().contains(searchTerm.toLowerCase())) {
+                matches.add(String.format("%s @ %s", name, func.getEntryPoint()));
+            }
+        }
+
+        Collections.sort(matches);
+
+        if (matches.isEmpty()) {
+            return "No functions matching '" + searchTerm + "'";
+        }
+        return paginateList(matches, offset, limit);
+    }
+}
