+++ src/main/java/com/lauriewired/util/ParseUtils.java
@@ -1,0 +1,119 @@
+package com.lauriewired.util;
+
+import com.sun.net.httpserver.HttpExchange;
+import ghidra.util.Msg;
+
+import java.io.IOException;
+import java.io.OutputStream;
+import java.net.URLDecoder;
+import java.nio.charset.StandardCharsets;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public final class ParseUtils {
+    /**
+     * Parse query parameters from the URL, e.g. ?offset=10&limit=100
+     */
+    public static Map<String, String> parseQueryParams(HttpExchange exchange) {
+        Map<String, String> result = new HashMap<>();
+        String query = exchange.getRequestURI().getQuery(); // e.g. offset=10&limit=100
+        if (query != null) {
+            String[] pairs = query.split("&");
+            for (String p : pairs) {
+                String[] kv = p.split("=");
+                if (kv.length == 2) {
+                    // URL decode parameter values
+                    try {
+                        String key = URLDecoder.decode(kv[0], StandardCharsets.UTF_8);
+                        String value = URLDecoder.decode(kv[1], StandardCharsets.UTF_8);
+                        result.put(key, value);
+                    } catch (Exception e) {
+                        Msg.error(ParseUtils.class, "Error decoding URL parameter", e);
+                    }
+                }
+            }
+        }
+        return result;
+    }
+
+    /**
+     * Parse post body form params, e.g. oldName=foo&newName=bar
+     */
+    public static Map<String, String> parsePostParams(HttpExchange exchange) throws IOException {
+        byte[] body = exchange.getRequestBody().readAllBytes();
+        String bodyStr = new String(body, StandardCharsets.UTF_8);
+        Map<String, String> params = new HashMap<>();
+        for (String pair : bodyStr.split("&")) {
+            String[] kv = pair.split("=");
+            if (kv.length == 2) {
+                // URL decode parameter values
+                try {
+                    String key = URLDecoder.decode(kv[0], StandardCharsets.UTF_8);
+                    String value = URLDecoder.decode(kv[1], StandardCharsets.UTF_8);
+                    params.put(key, value);
+                } catch (Exception e) {
+                    Msg.error(ParseUtils.class, "Error decoding URL parameter", e);
+                }
+            }
+        }
+        return params;
+    }
+
+    /**
+     * Convert a list of strings into one big newline-delimited string, applying offset & limit.
+     */
+    public static String paginateList(List<String> items, int offset, int limit) {
+        int start = Math.max(0, offset);
+        int end   = Math.min(items.size(), offset + limit);
+
+        if (start >= items.size()) {
+            return ""; // no items in range
+        }
+        List<String> sub = items.subList(start, end);
+        return String.join("\n", sub);
+    }
+
+    /**
+     * Parse an integer from a string, or return defaultValue if null/invalid.
+     */
+    public static int parseIntOrDefault(String val, int defaultValue) {
+        if (val == null) return defaultValue;
+        try {
+            return Integer.parseInt(val);
+        }
+        catch (NumberFormatException e) {
+            return defaultValue;
+        }
+    }
+
+    /**
+     * Escape non-ASCII chars to avoid potential decode issues.
+     */
+    public static String escapeNonAscii(String input) {
+        if (input == null) return "";
+        StringBuilder sb = new StringBuilder();
+        for (char c : input.toCharArray()) {
+            if (c >= 32 && c < 127) {
+                sb.append(c);
+            }
+            else {
+                sb.append("\\x");
+                sb.append(Integer.toHexString(c & 0xFF));
+            }
+        }
+        return sb.toString();
+    }
+
+    /**
+     * Send responses
+     */
+    public static void sendResponse(HttpExchange exchange, String response) throws IOException {
+        byte[] bytes = response.getBytes(StandardCharsets.UTF_8);
+        exchange.getResponseHeaders().set("Content-Type", "text/plain; charset=utf-8");
+        exchange.sendResponseHeaders(200, bytes.length);
+        try (OutputStream os = exchange.getResponseBody()) {
+            os.write(bytes);
+        }
+    }
+}
