+++ src/main/resources/extension.properties
@@ -2,5 +2,5 @@ name=GhidraMCP
 description=A plugin that runs an embedded HTTP server to expose program data.
 author=LaurieWired
 createdOn=2025-03-22
-version=11.3.1
-ghidraVersion=11.3.1
+version=11.3.2
+ghidraVersion=11.3.2
